+++ src/app/cmd/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod add_server;